requirements.txt
Brotli==1.0.9
click==8.0.1
cycler==0.10.0
dash==1.20.0
dash-core-components==1.16.0
dash-cytoscape==0.2.0
dash-html-components==1.1.3
dash-renderer==1.9.1
dash-table==4.11.3
decorator==4.4.2
Flask==2.0.1
Flask-Compress==1.9.0
future==0.18.2
imageio==2.9.0
itsdangerous==2.0.1
Jinja2==3.0.1
joblib==1.0.1
kiwisolver==1.3.1
MarkupSafe==2.0.1
matplotlib==3.4.1
networkx==2.5.1
numpy==1.20.2
pandas==1.2.4
Pillow==8.2.0
plotly==4.14.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2021.1
PyWavelets==1.1.1
retrying==1.3.3
scikit-image==0.18.1
scikit-learn==0.24.2
scipy==1.6.3
six==1.15.0
threadpoolctl==2.1.0
tifffile==2021.4.8
torch==1.8.1
torchvision==0.9.1
tqdm==4.60.0
typing-extensions==3.10.0.0
Werkzeug==2.0.1
